feat: implemented utils
